asdfdsaf

+++ src/Interface.py
@@ -48,7 +48,7 @@ class Gripper(Node):
     def gripper_open(self):
         try:
             msg = Float32MultiArray()
-            msg.data = [float(170)]
+            msg.data = [float(0)]
             self.publisher.publish(msg)
             self.get_logger().info("그리퍼 열기")
         except Exception as e:
@@ -57,7 +57,7 @@ class Gripper(Node):
     def gripper_close(self):
         try:
             msg = Float32MultiArray()
-            msg.data = [float(0)]
+            msg.data = [float(160)]
             self.publisher.publish(msg)
             self.get_logger().info("그리퍼 닫기")
         except Exception as e:
